requirements.txt
numpy
tensorboard
scikit-learn
matplotlib
opencv-python
PyYAML
--index-url https://download.pytorch.org/whl/cu118
torch
torchvision